move NewSessionReaction to core
