--- channels/jaicp/src/main/kotlin/com/justai/jaicf/channel/jaicp/reactions/reaction/NewSessionReaction.kt
@@ -1,9 +1,0 @@
-package com.justai.jaicf.channel.jaicp.reactions.reaction
-
-import com.justai.jaicf.logging.Reaction
-
-data class NewSessionReaction(
-    override val fromState: String
-) : Reaction(fromState) {
-    override fun toString(): String = """start new session from state $fromState"""
-}
